TASK: extract i18n logic from I18n component
This would allow to render translations into simple strings, for cases
when using React components is not possible (e.g. input placeholders).
